Extracted tasks out of MfFactionApplyCommand into separate classes to increase testability
